backend/main/requirements.txt
fastapi
uvicorn
google.generativeai